requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
alpha-vantage==2.3.1